Gunner doesnt pay me enough either

+++ kinetic/compute.py
@@ -3,7 +3,7 @@ import os, time, glob
 
 def compute(m, v):
     """Return filename of plot of the damped_vibration function."""
-    e=m*v*v/2
+    e=float(m)*float(v*v)/2
     return e
     # Use time since Jan 1, 1970 in filename in order make
     # a unique filename that the browser has not chached
